requirements.txt
Django>=4.2.0
Pillow>=9.0.0
qrcode[pil]>=7.3.0
opencv-python>=4.5.0
numpy>=1.21.0
face-recognition>=1.3.0
